Fix model

+++ db/migrate/20160817210326_remove_estado_colmena_procesalfrom_case.rb
@@ -1,5 +1,7 @@
 class RemoveEstadoColmenaProcesalfromCase < ActiveRecord::Migration[5.0]
 	def change
-		remove_column :cases, :estado_colmena_procesal
+		unless table_exists? :cases, :estado_colmena_procesal
+			remove_column :cases, :estado_colmena_procesal
+		end
 	end
 end
